requirements.txt
fastapi
uvicorn
torch
numpy
opencv-python
facenet-pytorch
torchvision
scipy
supabase
python-dotenv
pillow
